@@ -44,58 +44,47 @@
     "compile": "tsc",
     "clean": "rm -rf {.nyc_output,coverage,lib} && git checkout -- lib",
     "dev": "tsc -w",
-    "lint-cov": "node bin/lint-cov",
-    "ncu": "ncu -a",
-    "nyc": "nyc",
-    "release-it": "release-it",
-    "test": "globstar -- _mocha \"test/spec/**/*.coffee\"",
-    "test-cov": "nyc -r lcov -r text -x test -x **/node_modules/** npm run -s test",
-    "test-ci-build": "node bin/test-ci-build",
-    "tsc": "tsc",
-    "vile": "vile a -u -n -x src.ts:lib.js"
+    "lint": "eslint .; coffeelint -q test",
+    "lint-cov": "nyc check-coverage --statements 100 --functions 100 --branches 100 --lines 100",
+    "release": "release-it",
+    "test": "mocha --require coffeescript/register -R dot -s 2000 --recursive 'test/spec/**/*.coffee'",
+    "test-cov": "nyc -r lcov -r text -x test/fixtures npm run test",
+    "test-ci-build": "npm run lint && npm run compile && npm run test-cov && npm run lint-cov",
+    "tsc": "tsc"
   },
   "dependencies": {
-    "@types/commander": "^2.9.1",
-    "@types/escodegen": "0.0.6",
-    "@types/esprima": "^4.0.1",
-    "@types/estraverse": "0.0.6",
-    "@types/estree": "0.0.38",
-    "@types/lodash": "^4.14.80",
-    "@types/node": "^10.5.7",
-    "cardinal": "^1.0.0",
-    "chalk": "^2.1.0",
-    "commander": "^2.10.0",
-    "escodegen": "^1.8.1",
-    "esprima": "^4.0.0",
-    "estraverse": "^4.2.0",
-    "lodash": "^4.17.4",
-    "typescript": "^2.3.4",
-    "walk-sync": "^0.3.2"
+    "@javascript-obfuscator/escodegen": "^2.3.0",
+    "@types/escodegen": "0.0.10",
+    "@types/esprima": "^4.0.6",
+    "@types/estraverse": "^5.1.7",
+    "@types/estree": "^1.0.5",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.10.5",
+    "cardinal": "^2.1.1",
+    "chalk": "^4.1.2",
+    "commander": "^11.1.0",
+    "espree": "^9.6.1",
+    "estraverse": "^5.3.0",
+    "lodash": "^4.17.21",
+    "typescript": "^5.3.3",
+    "walk-sync": "^3.0.0"
   },
   "devDependencies": {
-    "chai": "^4.0.2",
-    "coffee-script": "^1.12.6",
-    "coffeelint": "^2.0.6",
-    "globstar": "^1.0.0",
-    "mimus": "^0.2.7",
-    "mocha": "^4.0.1",
-    "npm-check-updates": "^2.12.0",
-    "nyc": "^11.3.0",
-    "release-it": "^3.1.2",
-    "retire": "^1.3.2",
-    "shelljs": "^0.7.8",
-    "sinon": "^4.0.2",
-    "sinon-chai": "^2.11.0",
-    "tslint": "^5.6.0",
-    "vile": "^0.19.0",
-    "vile-coffeelint": "^0.1.16",
-    "vile-depcheck": "^0.1.13",
-    "vile-eclint": "^0.1.14",
-    "vile-escomplex": "^0.1.5",
-    "vile-git": "^0.5.6",
-    "vile-nsp": "^0.2.14",
-    "vile-retire": "^0.1.3",
-    "vile-synt": "^0.1.2",
-    "vile-tslint": "^0.3.3"
+    "@stylistic/eslint-plugin-ts": "^1.5.1",
+    "@typescript-eslint/eslint-plugin": "^6.15.0",
+    "@typescript-eslint/parser": "^6.15.0",
+    "chai": "^4.3.10",
+    "coffeelint": "^2.1.0",
+    "coffeescript": "^2.7.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-jsdoc": "^46.9.1",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "mocha": "^10.2.0",
+    "npm-check-updates": "^16.14.12",
+    "nyc": "^15.1.0",
+    "release-it": "^17.0.1",
+    "retire": "^4.3.4",
+    "sinon": "^17.0.1",
+    "sinon-chai": "^3.7.0"
   }
 }
