@@ -1,22 +1,25 @@
-import program = require("commander")
-import similar = require("./similar")
-import fs_collector = require("./cli/file_collector")
+import { program } from "commander"
+import * as espree from 'espree'
+import * as similar from "./similar"
+import * as fs_collector from "./cli/file_collector"
 
-const pkg = require("./../package.json")
+const pkg = require("./../package.json") as synt.Package
 
 const compare = (
   targets : string[],
   opts : synt.CLIOptions
 ) => {
   const files = fs_collector.files(targets)
-  const nocolors : boolean = !!opts.disablecolors
+  const ecmaVersion = opts.ecmaVersion && opts.ecmaVersion != "latest" ?
+    parseInt(opts.ecmaVersion) : opts.ecmaVersion
+  const { js, ts } = similar.compare(files, { ...opts, ecmaVersion })
 
-  fs_collector.print(files, nocolors)
+  similar.print(js, opts.color)
+  similar.print(ts, opts.color)
 
-  const { js, ts } = similar.compare(files, opts)
-
-  similar.print(js, nocolors)
-  similar.print(ts, nocolors)
+  if (opts.exitCode && (Object.keys(js).length > 0 || Object.keys(ts).length > 0)) {
+    process.exit(1)
+  }
 }
 
 const configure = () => {
@@ -29,17 +32,21 @@ const configure = () => {
       `Lowest % similarity to look for ` +
       `[default=${ similar.DEFAULT_THRESHOLD }].`)
     .option(
-      "-m, --minlength [number]",
+      "-m, --min-length [number]",
       `Default token length a function needs to be to compare it ` +
       `[default=${ similar.DEFAULT_TOKEN_LENGTH }].`)
     .option(
-      "-n, --ngram [number]",
+      "-g, --ngram [number]",
       `Specify ngram length for comparing token sequences. ` +
       `[default=${ similar.DEFAULT_NGRAM_LENGTH },2,3...]`)
-    .option("-d, --disablecolors", "Disable color output")
+    .option("-n, --no-color", "Disable color output")
+    .option("-e, --exit-code", "Exit with a nonzero code when issues found")
     .option(
-      "--estype [value]",
-      "Set the JavaScript parser source type [default=module,script]")
+      "-t, --source-type [value]",
+      "Set JS source type [default=module,script,commonjs]")
+    .option(
+      "-a, --ecma-version [value]",
+      `Set JS version [default=latest,${espree.supportedEcmaVersions.join(",")}]`)
     .action(compare)
 
   program.on("--help", () => {
@@ -60,4 +67,4 @@ const interpret = (argv : string[]) => {
   program.parse(argv)
 }
 
-export = { interpret }
+export { interpret }
