@@ -1,10 +1,7 @@
-import fs = require("fs")
-import path = require("path")
-import _ = require("lodash")
+import * as fs from "fs"
+import * as path from "path"
+import * as _ from "lodash"
 import walk_sync = require("walk-sync")
-
-// HACK: chalk types don't support import?
-const chalk = require("chalk")
 
 const all_files = (target : string) : string[] => {
   if (fs.statSync(target).isDirectory()) {
@@ -20,7 +17,8 @@ const normalize_cli_targets = (
 ) : string[] => {
   targets = _.concat([], targets)
 
-  let files = _.uniq(_.reduce(targets, (paths, target) =>
+  let files : string[] = _.uniq(_.reduce(targets,
+    (paths : string[], target : string) =>
       _.concat(paths, all_files(target)), []))
 
   files = _.map(files, (file : string) =>
@@ -30,19 +28,6 @@ const normalize_cli_targets = (
     /\.(js|ts)$/.test(file))
 }
 
-const print_found = (
-  files : string[],
-  nocolors : boolean
-) =>
-  _.each(files, (file) => {
-    if (nocolors) {
-      console.log("found:", file)
-    } else {
-      console.log(chalk.gray("found:"), chalk.green(file))
-    }
-  })
-
-export = {
-  files: normalize_cli_targets,
-  print: print_found
+export {
+  normalize_cli_targets as files
 }
