@@ -1,4 +1,4 @@
 "use strict";
-var similar = require("./similar");
-var library = similar;
+const similar = require("./similar");
+const library = similar;
 module.exports = library;
