@@ -1,95 +1,89 @@
 "use strict";
-var _ = require("lodash");
-var ngram = require("./similar/ngram");
-var parse_js = require("./similar/javascript");
-var parse_ts = require("./similar/typescript");
-var print = require("./similar/print");
-var DEFAULT_NGRAM_LENGTH = 1;
-var DEFAULT_THRESHOLD = 70;
-var DEFAULT_TOKEN_LENGTH = 10;
-var similarity = function (src, cmp) {
-    var a = _.uniq(src);
-    var b = _.uniq(cmp);
-    var i = _.intersection(a, b);
-    var u = _.union(a, b);
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.print = exports.compare = exports.DEFAULT_TOKEN_LENGTH = exports.DEFAULT_THRESHOLD = exports.DEFAULT_NGRAM_LENGTH = void 0;
+const _ = require("lodash");
+const ngram = require("./similar/ngram");
+const parse_js = require("./similar/javascript");
+const parse_ts = require("./similar/typescript");
+const print_1 = require("./similar/print");
+Object.defineProperty(exports, "print", { enumerable: true, get: function () { return print_1.print; } });
+const DEFAULT_NGRAM_LENGTH = 1;
+exports.DEFAULT_NGRAM_LENGTH = DEFAULT_NGRAM_LENGTH;
+const DEFAULT_THRESHOLD = 70;
+exports.DEFAULT_THRESHOLD = DEFAULT_THRESHOLD;
+const DEFAULT_TOKEN_LENGTH = 10;
+exports.DEFAULT_TOKEN_LENGTH = DEFAULT_TOKEN_LENGTH;
+const similarity = (src, cmp) => {
+    const a = _.uniq(src);
+    const b = _.uniq(cmp);
+    const i = _.intersection(a, b);
+    const u = _.union(a, b);
     return _.toNumber(_.toNumber((i.length / u.length) * 100)
         .toFixed(0));
 };
-var parse_token_length = function (str) {
-    return _.isEmpty(str) ?
-        DEFAULT_TOKEN_LENGTH :
-        _.toNumber(str);
-};
-var parse_ngram_length = function (str) {
-    return _.isEmpty(str) ?
-        DEFAULT_NGRAM_LENGTH :
-        _.toNumber(str);
-};
-var parse_threshold = function (str) {
-    var threshold = _.toNumber(str);
+const parse_token_length = (str) => _.isEmpty(str) ?
+    DEFAULT_TOKEN_LENGTH :
+    _.toNumber(str);
+const parse_ngram_length = (str) => _.isEmpty(str) ?
+    DEFAULT_NGRAM_LENGTH :
+    _.toNumber(str);
+const parse_threshold = (str) => {
+    const threshold = _.toNumber(str);
     return threshold || DEFAULT_THRESHOLD;
 };
-var is_ts_ancestor = function (src, cmp) {
-    var match = false;
-    var last = src.ast;
+const is_ts_ancestor = (src, cmp) => {
+    let match = false;
+    let last = src.ast;
     while (true) {
-        var parent_1 = last.parent;
-        if (parent_1 === cmp.ast)
+        const { parent } = last;
+        if (parent === cmp.ast)
             match = true;
-        if (!parent_1 || last === parent_1 || match)
+        if (!parent || last === parent || match)
             break;
-        last = parent_1;
+        last = parent;
     }
     return match;
 };
-var false_positive = function (src, cmp, t_len) {
-    var same_node = function () { return cmp.ast === src.ast; };
-    var size_is_too_different = function () {
-        var l1 = src.tokens.length;
-        var l2 = cmp.tokens.length;
+const false_positive = (src, cmp, t_len) => {
+    const same_node = () => cmp.ast === src.ast;
+    const size_is_too_different = () => {
+        const l1 = src.tokens.length;
+        const l2 = cmp.tokens.length;
         return l1 * 2 < l2 || l2 * 2 < l1;
     };
-    var one_is_too_short = function () {
-        return src.tokens.length < t_len ||
-            cmp.tokens.length < t_len;
-    };
-    var subset_of_other = function () {
-        var is_eithers_ancestor = function () {
-            return is_ts_ancestor(src, cmp) ||
-                is_ts_ancestor(cmp, src);
-        };
-        var is_eithers_middle = function () {
-            var src_j = src.tokens.join("");
-            var cmp_j = cmp.tokens.join("");
+    const one_is_too_short = () => src.tokens.length < t_len ||
+        cmp.tokens.length < t_len;
+    const subset_of_other = () => {
+        const is_eithers_ancestor = () => is_ts_ancestor(src, cmp) ||
+            is_ts_ancestor(cmp, src);
+        const is_eithers_middle = () => {
+            const src_j = src.tokens.join("");
+            const cmp_j = cmp.tokens.join("");
             return src_j !== cmp_j &&
                 (_.includes(src_j, cmp_j) ||
                     _.includes(cmp_j, src_j));
         };
         return is_eithers_ancestor() || is_eithers_middle();
     };
-    var both_are_not_classes = function () {
-        return (src.is_class && !cmp.is_class) ||
-            (!src.is_class && cmp.is_class);
-    };
+    const both_are_not_classes = () => (src.is_class && !cmp.is_class) ||
+        (!src.is_class && cmp.is_class);
     return same_node() ||
         both_are_not_classes() ||
         subset_of_other() ||
         one_is_too_short() ||
         size_is_too_different();
 };
-var each_pair = function (items, callback) {
-    _.each(items, function (src) {
-        _.each(items, function (cmp) {
+const each_pair = (items, callback) => {
+    _.each(items, (src) => {
+        _.each(items, (cmp) => {
             callback(src, cmp);
         });
     });
 };
-var filter_redundencies = function (group) {
-    _.each(group, function (results, sim) {
-        group[sim] = _.reduce(results, function (new_arr, result) {
-            var already_added = _.some(new_arr, function (result_two) {
-                return _.xor(result, result_two).length === 0;
-            });
+const filter_redundencies = (group) => {
+    _.each(group, (results, sim) => {
+        group[sim] = _.reduce(results, (new_arr, result) => {
+            const already_added = _.some(new_arr, (result_two) => _.xor(result, result_two).length === 0);
             if (!already_added) {
                 new_arr.push(result);
             }
@@ -98,22 +92,22 @@ var filter_redundencies = function (group) {
     });
     return group;
 };
-var _compare = function (files, ftype, opts) {
-    var is_ts = ftype === "ts";
-    var is_file = is_ts ? /\.ts$/ : /\.js$/;
-    files = _.filter(files, function (file) { return is_file.test(file); });
-    var parse = is_ts ? parse_ts : parse_js;
-    var items = parse.find(files, opts);
-    var group = {};
-    var t_len = parse_token_length(_.toString(opts.minlength));
-    var n_len = parse_ngram_length(_.toString(opts.ngram));
-    var sim_min = parse_threshold(_.toString(opts.similarity));
-    each_pair(items, function (src, cmp) {
+const _compare = (files, ftype, opts) => {
+    const is_ts = ftype === "ts";
+    const is_file = is_ts ? /\.ts$/ : /\.js$/;
+    files = _.filter(files, (file) => is_file.test(file));
+    const parse = is_ts ? parse_ts : parse_js;
+    const items = parse.find(files, opts);
+    const group = {};
+    const t_len = parse_token_length(_.toString(opts.minLength));
+    const n_len = parse_ngram_length(_.toString(opts.ngram));
+    const sim_min = parse_threshold(_.toString(opts.similarity));
+    each_pair(items, (src, cmp) => {
         if (false_positive(src, cmp, t_len))
             return;
-        var src_grams = ngram.generate(src.tokens, n_len);
-        var cmp_grams = ngram.generate(cmp.tokens, n_len);
-        var val = similarity(src_grams, cmp_grams);
+        const src_grams = ngram.generate(src.tokens, n_len);
+        const cmp_grams = ngram.generate(cmp.tokens, n_len);
+        const val = similarity(src_grams, cmp_grams);
         if (val < sim_min)
             return;
         if (_.isEmpty(group[val]))
@@ -122,17 +116,10 @@ var _compare = function (files, ftype, opts) {
     });
     return filter_redundencies(group);
 };
-var compare = function (files, opts) {
-    if (opts === void 0) { opts = {}; }
+const compare = (files, opts = {}) => {
     files = _.concat([], files);
-    var js_group = _compare(files, "js", opts);
-    var ts_group = _compare(files, "ts", opts);
+    const js_group = _compare(files, "js", opts);
+    const ts_group = _compare(files, "ts", opts);
     return { js: js_group, ts: ts_group };
 };
-module.exports = {
-    DEFAULT_NGRAM_LENGTH: DEFAULT_NGRAM_LENGTH,
-    DEFAULT_THRESHOLD: DEFAULT_THRESHOLD,
-    DEFAULT_TOKEN_LENGTH: DEFAULT_TOKEN_LENGTH,
-    compare: compare,
-    print: print.print
-};
+exports.compare = compare;
