@@ -1,52 +1,57 @@
 "use strict";
-var fs = require("fs");
-var _ = require("lodash");
-var codegen = require("escodegen");
-var esprima = require("esprima");
-var estraverse = require("estraverse");
-var FUNCTION_OR_CLASS_NODE = [
-    esprima.Syntax.ArrowFunctionExpression,
-    esprima.Syntax.ClassDeclaration,
-    esprima.Syntax.FunctionDeclaration,
-    esprima.Syntax.FunctionExpression
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.find = void 0;
+const escodegen = require("@javascript-obfuscator/escodegen");
+const espree = require("espree");
+const estraverse = require("estraverse");
+const fs = require("fs");
+const _ = require("lodash");
+const FUNCTION_OR_CLASS_NODE = [
+    espree.Syntax.ArrowFunctionExpression,
+    espree.Syntax.ClassDeclaration,
+    espree.Syntax.FunctionDeclaration,
+    espree.Syntax.FunctionExpression
 ];
-var normalize = function (token_list) {
-    return _.map(token_list, function (t) { return t.value; });
+const normalize = (token_list) => _.map(token_list, (t) => t.value);
+const tokenize = (code, opts) => {
+    const tokenize_opts = {
+        ecmaVersion: opts.ecmaVersion || "latest",
+        sourceType: opts.sourceType || "module"
+    };
+    return normalize(espree.tokenize(code, tokenize_opts));
 };
-var tokenize = function (code) {
-    return normalize(esprima.tokenize(code));
+const astify = (code, opts) => {
+    const module_type = opts.sourceType || "module";
+    const ecma_version = opts.ecmaVersion || "latest";
+    const parse_opts = {
+        ecmaVersion: ecma_version,
+        loc: true,
+        sourceType: module_type
+    };
+    return espree.parse(code, parse_opts);
 };
-var astify = function (code, opts) {
-    var parse_opts = { loc: true };
-    var method = _.get(opts, "estype", "module");
-    return method == "script" ?
-        esprima.parseScript(code, parse_opts) :
-        esprima.parseModule(code, parse_opts);
+const ast_to_code = (node) => {
+    const opts = { format: { indent: { style: "  " } } };
+    return escodegen.generate(node, opts);
 };
-var ast_to_code = function (node) {
-    var opts = { format: { indent: { style: "  " } } };
-    return codegen.generate(node, opts);
-};
-var is_a_method_or_class = function (node) {
-    return _.some(FUNCTION_OR_CLASS_NODE, function (type) { return type === node.type; });
-};
-var line_info = function (node) { return node.loc; };
-var parse_methods_and_classes = function (root_node, filepath) {
-    var entries = [];
+const is_a_method_or_class = (node) => _.some(FUNCTION_OR_CLASS_NODE, (type) => type === node.type);
+const line_info = (node) => node.loc;
+const parse_methods_and_classes = (root_node, filepath, opts) => {
+    const entries = [];
     estraverse.traverse(root_node, {
-        enter: function (node, parent) {
+        enter(node, _parent) {
             if (!is_a_method_or_class(node))
                 return;
-            var method = ast_to_code(node);
-            var tokens = tokenize(method);
-            var is_class = node.type === esprima.Syntax.ClassDeclaration;
-            var result = {
+            const method = ast_to_code(node);
+            const tokens = tokenize(method, opts);
+            const is_class = node.type === espree.Syntax.ClassDeclaration;
+            const result = {
                 ast: node,
                 code: method,
-                is_class: is_class,
+                is_class,
                 path: filepath,
                 pos: line_info(node),
-                tokens: tokens,
+                tokens,
                 type: node.type
             };
             entries.push(result);
@@ -54,19 +59,15 @@ var parse_methods_and_classes = function (root_node, filepath) {
     });
     return entries;
 };
-var find_similar_methods_and_classes = function (filepaths, opts) {
-    return _.flatMap(filepaths, function (filepath) {
-        var code = fs.readFileSync(filepath).toString();
-        var node;
-        try {
-            node = astify(code, opts);
-        }
-        catch (err) {
-            throw new Error("in " + filepath + "\n\n" + err.stack);
-        }
-        return parse_methods_and_classes(node, filepath);
-    });
-};
-module.exports = {
-    find: find_similar_methods_and_classes
-};
+const find_similar_methods_and_classes = (filepaths, opts) => _.flatMap(filepaths, (filepath) => {
+    const code = fs.readFileSync(filepath).toString();
+    let node;
+    try {
+        node = astify(code, opts);
+    }
+    catch (err) {
+        throw new Error(`in ${filepath}\n\n${_.get(err, "stack", err)}`);
+    }
+    return parse_methods_and_classes(node, filepath, opts);
+});
+exports.find = find_similar_methods_and_classes;
