@@ -1,17 +1,18 @@
 "use strict";
-var generate = function (arr, len) {
-    if (len === void 0) { len = 1; }
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.generate = void 0;
+const generate = (arr, len = 1) => {
     if (len > arr.length)
         len = 1;
     if (len == 1)
         return arr;
-    var sets = [];
-    arr.forEach(function (token, index) {
-        var s_len = index + len;
+    const sets = [];
+    arr.forEach((token, index) => {
+        const s_len = index + len;
         if (s_len <= arr.length) {
             sets.push(arr.slice(index, s_len).join(""));
         }
     });
     return sets;
 };
-module.exports = { generate: generate };
+exports.generate = generate;
