@@ -1,56 +1,62 @@
 "use strict";
-var _ = require("lodash");
-var cardinal = require("cardinal");
-var chalk = require("chalk");
-var print = function (group, nocolors) {
-    _.each(group, function (results, sim) {
-        _.each(results, function (result) {
-            var src = result[0], cmp = result[1];
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.print = void 0;
+const _ = require("lodash");
+const cardinal = require("cardinal");
+const chalk = require("chalk");
+const print = (group, color) => {
+    _.each(group, (results, sim) => {
+        _.each(results, (result) => {
+            const [src, cmp] = result;
             console.log("");
-            var match_sim = sim + "% match";
-            if (nocolors) {
+            const match_sim = sim + "% similar";
+            if (color) {
+                console.log(chalk.red.bold(match_sim));
+            }
+            else {
                 console.log(match_sim);
             }
-            else {
-                console.log(chalk.white.bgRed.bold(match_sim));
-            }
             console.log("");
-            if (nocolors) {
-                console.log("in: " + src.path);
-            }
-            else {
+            if (color) {
                 console.log(chalk.gray("in: ") + chalk.green(src.path));
             }
-            console.log("");
-            if (nocolors) {
-                console.log(src.code);
-            }
             else {
+                console.log(`in: ${src.path}`);
+            }
+            console.log("");
+            if (color) {
                 console.log(cardinal.highlight(src.code, {
                     firstline: src.pos.start.line,
                     linenos: true
                 }));
             }
+            else {
+                console.log(src.code.split("\n").map((line, idx) => {
+                    return `${src.pos.start.line + idx + 1}: ${line}`;
+                }).join("\n"));
+            }
             console.log("");
             if (src.path !== cmp.path) {
-                if (nocolors) {
-                    console.log("in: " + cmp.path);
+                if (color) {
+                    console.log(chalk.gray("in: ") + chalk.green(cmp.path));
                 }
                 else {
-                    console.log(chalk.gray("in: ") + chalk.green(cmp.path));
+                    console.log(`in: ${cmp.path}`);
                 }
                 console.log("");
             }
-            if (nocolors) {
-                console.log(cmp.code);
-            }
-            else {
+            if (color) {
                 console.log(cardinal.highlight(cmp.code, {
                     firstline: cmp.pos.start.line,
                     linenos: true
                 }));
             }
+            else {
+                console.log(cmp.code.split("\n").map((line, idx) => {
+                    return `${cmp.pos.start.line + idx + 1}: ${line}`;
+                }).join("\n"));
+            }
         });
     });
 };
-module.exports = { print: print };
+exports.print = print;
