@@ -1,13 +1,15 @@
 "use strict";
-var _ = require("lodash");
-var fs = require("fs");
-var ts = require("typescript");
-var STOP_AT_NODES = [
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.find = void 0;
+const _ = require("lodash");
+const fs = require("fs");
+const ts = require("typescript");
+const STOP_AT_NODES = [
     ts.SyntaxKind.ArrayType,
     ts.SyntaxKind.PrefixUnaryExpression,
     ts.SyntaxKind.RegularExpressionLiteral
 ];
-var FUNCTION_OR_CLASS_NODE = [
+const FUNCTION_OR_CLASS_NODE = [
     ts.SyntaxKind.ArrowFunction,
     ts.SyntaxKind.ClassDeclaration,
     ts.SyntaxKind.Constructor,
@@ -15,7 +17,7 @@ var FUNCTION_OR_CLASS_NODE = [
     ts.SyntaxKind.FunctionExpression,
     ts.SyntaxKind.MethodDeclaration
 ];
-var PASSTHROUGH_NODES = FUNCTION_OR_CLASS_NODE.concat([
+const PASSTHROUGH_NODES = FUNCTION_OR_CLASS_NODE.concat([
     ts.SyntaxKind.ArrowFunction,
     ts.SyntaxKind.ArrayLiteralExpression,
     ts.SyntaxKind.BinaryExpression,
@@ -48,20 +50,14 @@ var PASSTHROUGH_NODES = FUNCTION_OR_CLASS_NODE.concat([
     ts.SyntaxKind.VariableDeclarationList,
     ts.SyntaxKind.VariableStatement
 ]);
-var astify = function (code, filepath, opts) {
+const astify = (code, filepath, _opts) => {
     ts.createProgram([filepath], {});
     return ts.createSourceFile(filepath, code, ts.ScriptTarget.Latest, true, ts.ScriptKind.TS);
 };
-var is_a_base_node = function (node) {
-    return _.some(STOP_AT_NODES, function (kind) { return kind === node.kind; });
-};
-var is_a_passthrough_node = function (node) {
-    return _.some(PASSTHROUGH_NODES, function (kind) { return kind === node.kind; });
-};
-var is_a_method_or_class = function (node) {
-    return _.some(FUNCTION_OR_CLASS_NODE, function (kind) { return kind === node.kind; });
-};
-var _tokenize = function (tokens) { return function (node) {
+const is_a_base_node = (node) => _.some(STOP_AT_NODES, (kind) => kind === node.kind);
+const is_a_passthrough_node = (node) => _.some(PASSTHROUGH_NODES, (kind) => kind === node.kind);
+const is_a_method_or_class = (node) => _.some(FUNCTION_OR_CLASS_NODE, (kind) => kind === node.kind);
+const _tokenize = (tokens) => (node) => {
     if (is_a_base_node(node)) {
         tokens.push(node.getText());
     }
@@ -72,21 +68,21 @@ var _tokenize = function (tokens) { return function (node) {
         tokens.push(node.getText());
         ts.forEachChild(node, _tokenize(tokens));
     }
-}; };
-var tokenize = function (ast_node) {
-    var tokens = [];
+};
+const tokenize = (ast_node) => {
+    const tokens = [];
     _tokenize(tokens)(ast_node);
     return tokens;
 };
-var find_all_methods_and_classes = function (list) { return function (node) {
+const find_all_methods_and_classes = (list) => (node) => {
     if (is_a_method_or_class(node)) {
         list.push(node);
     }
     ts.forEachChild(node, find_all_methods_and_classes(list));
-}; };
-var line_info = function (node, root_node) {
-    var _a = root_node
-        .getLineAndCharacterOfPosition(node.getStart()), line = _a.line, character = _a.character;
+};
+const line_info = (node, root_node) => {
+    const { line, character } = root_node
+        .getLineAndCharacterOfPosition(node.getStart());
     return {
         start: {
             column: character,
@@ -94,32 +90,28 @@ var line_info = function (node, root_node) {
         }
     };
 };
-var parse_methods_and_classes = function (node, root_node, path) {
-    var methods_and_classes = [];
+const parse_methods_and_classes = (node, root_node, path) => {
+    const methods_and_classes = [];
     find_all_methods_and_classes(methods_and_classes)(node);
-    return _.map(methods_and_classes, function (method_or_class) {
-        var tokens = tokenize(method_or_class);
-        var code = method_or_class.getText();
-        var is_class = method_or_class.kind == ts.SyntaxKind.ClassDeclaration;
+    return _.map(methods_and_classes, (method_or_class) => {
+        const tokens = tokenize(method_or_class);
+        const code = method_or_class.getText();
+        const is_class = method_or_class.kind == ts.SyntaxKind.ClassDeclaration;
         return {
             ast: method_or_class,
-            code: code,
-            is_class: is_class,
-            path: path,
+            code,
+            is_class,
+            path,
             pos: line_info(method_or_class, root_node),
-            tokens: tokens,
+            tokens,
             type: ts.SyntaxKind[method_or_class.kind]
         };
     });
 };
-var find_similar_methods_and_classes = function (filepaths, opts) {
-    return _.flatMap(filepaths, function (filepath) {
-        var code = fs.readFileSync(filepath).toString();
-        var node = astify(code, filepath, opts);
-        var root_node = node;
-        return parse_methods_and_classes(node, root_node, filepath);
-    });
-};
-module.exports = {
-    find: find_similar_methods_and_classes
-};
+const find_similar_methods_and_classes = (filepaths, opts) => _.flatMap(filepaths, (filepath) => {
+    const code = fs.readFileSync(filepath).toString();
+    const node = astify(code, filepath, opts);
+    const root_node = node;
+    return parse_methods_and_classes(node, root_node, filepath);
+});
+exports.find = find_similar_methods_and_classes;
