@@ -1,11 +1,12 @@
 declare module "cardinal" {
   export interface CardinalOptions {
-    linenos?    : boolean;
-    firstline? : number;
+    linenos   ?: boolean
+    firstline ?: number
+    theme     ?: string
   }
 
   function highlight(
-    code : string,
-    opts? : CardinalOptions
-  ) : string;
+    code  : string,
+    opts ?: CardinalOptions
+  ) : string
 }
