@@ -1,12 +1,12 @@
 declare module "walk-sync" {
   interface WalkSyncOptions {
-    directories? : boolean;
+    directories ?: boolean
   }
 
   function walk_sync(
     target : string,
-    opts? : WalkSyncOptions
-  ) : string[];
+    opts   ?: WalkSyncOptions
+  ) : string[]
 
   export = walk_sync
 }
